+++ docs/changelog.md
@@ -2,6 +2,7 @@
 ## Version 0.2.1
 
 - Add Stop/Limit on trades
+- Previous indicator skip empty frames
 
 ## Version 0.2.0